Upgrade the proto dependency to 30.2 and drop C++ AWS KMS support in the cross language tests.
AWS KMS includes curl as a dependency, which in turn uses https://github.com/curl/curl/blob/master/lib/vtls/openssl.c. This makes it more difficult to build in the same binary as Tink.

Since we seem to be heading to drop AWS KMS support, I will remove it from the cross language tests; if we keep supporting it we will have to add it back.

PiperOrigin-RevId: 779982816
Change-Id: Id123ba9f1180b0c8d8cb5e35e806db4cf75613f2